Possible solution for #37 and #60
Problem: If a javascript file's last line is a comment, Cordova's wrapping mechanism might fail to produce syntactically correct AMD modules. This was happening to the underscore lib, since it's last line is a source map definition which eventually broke the wrapping in some cases (not yet known what else is needed to trigger this problem).

Solution: Removed the commented last line of Underscore.js to battle the misbehaving wrapping mechanism. This might fix the issues when people are experiencing that the plugin is breaking their apps simply because of being added to the project.